Fix multiple issues: QR warnings, admin tickets, error modal
- Fix QRService float to int conversion warnings (cast to int)
- Fix admin/tickets template to use correct field names (code, attendee_name, attendee_company)
- Add attractive error modal instead of browser alert() for registration errors
- Modal shows friendly messages for common errors (already registered, session expired, sold out)
- Modal supports keyboard (Escape) and overlay click to close

+++ src/Services/QRService.php
@@ -136,8 +136,8 @@ class QRService
         $fontSize = 3;
         $text = substr($data, 0, 20);
         $textWidth = strlen($text) * imagefontwidth($fontSize);
-        $x = ($this->size - $textWidth) / 2;
-        $y = ($this->size - imagefontheight($fontSize)) / 2;
+        $x = (int)(($this->size - $textWidth) / 2);
+        $y = (int)(($this->size - imagefontheight($fontSize)) / 2);
 
         imagestring($image, $fontSize, $x, $y - 20, 'QR Code', $black);
         imagestring($image, $fontSize, $x, $y + 10, $text, $black);
